src/pyproject.toml
[tool.poetry]
name = "panel"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.6"
django-rq = "^2.4.0"
djangorestframework = "^3.12.2"
Hypercorn = "^0.11.2"
"luma.led_matrix" = "^1.5.0"
luma = "^2.1.3"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.11b1"
isort = "^5.10.1"
flake8 = "^4.0.1"
pydantic = "^1.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
